Portal and explosions
Only two portals appear, the time the portal is prevented to be reused is added as param in types. Portal tiles are defined in level generators and portal dynamics in the forward model Three different geometry for the explosions are used: normal, diagonal and circle of fire. They appear randomly, the normal with 0.5 probability, the others with 0.25. All refering to explosions is defined in the bomb script.